created a student context for state to prevent prop drilling and mixing formJson and inidividual states for each portion of the form. Began modularizing into smaller components starting with StudentForm.tsx
