+++ src/pages/students/index.tsx
@@ -23,9 +23,20 @@ import StudentsDetailsTable from './StudentsDetailsTable';
 import StudentUploader from './StudentUploader';
 import { RefreshContext, useNotifications } from '@digitalaidseattle/core';
 import FailedStudentsModal from './FailedStudentsModal';
-import { FailedStudent, Student, StudentField, TimeWindow } from '../../api/types';
+import { FailedStudent, Student, TimeWindow } from '../../api/types';
 import AddStudentModal from './AddStudentModal';
 import { studentService } from '../../api/ceStudentService';
+import { createContext } from 'react';
+
+interface StudentContextType {
+    student: Student,
+    setStudent: React.Dispatch<React.SetStateAction<Student>>
+}
+
+export const StudentContext = createContext<StudentContextType>({
+    student: {} as Student,
+    setStudent: () => {}
+})
 
 const UploadSection = () => {
     const notifications = useNotifications();
@@ -35,14 +46,6 @@ const UploadSection = () => {
     const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
     const [isAddStudentModalOpen, setIsAddStudentModalOpen] = useState<boolean>(false)
     const [availabilities, setAvailabilities] = useState<TimeWindow[]>([])
-
-    const studentField: StudentField[] = [
-        { key: 'name', label: 'Full Name', type: 'string', required: true },
-        { key: 'age', label: 'Age', type: 'number', required: true },
-        { key: 'email', label: 'Email', type: 'email', required: true },
-        { key: 'country', label: 'Country', type: 'string', required: true },
-    ];
-
 
     const handleUpdate = (resp: any) => {
         setRefresh(refresh + 1);
@@ -113,7 +116,6 @@ const UploadSection = () => {
                 isAddStudentModalOpen={isAddStudentModalOpen}
                 onClose={() => handleCloseAddStudentModal()}
                 handleAddStudent={handleAddStudent}
-                studentField={studentField}
                 availabilities={availabilities}
                 setAvailabilities={setAvailabilities}
             />
@@ -121,11 +123,14 @@ const UploadSection = () => {
     )
 }
 const StudentsPage: React.FC = () => {
+    const [student, setStudent] = useState<Student>({} as Student);
     return (
-        <MainCard title="Students Page">
-            <UploadSection />
-            <StudentsDetailsTable />
-        </MainCard>
+        <StudentContext.Provider value={{student, setStudent}}>
+            <MainCard title="Students Page">
+                <UploadSection />
+                <StudentsDetailsTable />
+            </MainCard>
+        </StudentContext.Provider>
     )
 };
